requirements.txt
alembic==1.9.4
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
omymodels==0.13.0
parsimonious==0.8.1
pip-chill==1.0.1
ply==3.11
psycopg2-binary==2.9.5
py-models-parser==0.6.0
pydantic==1.10.5
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.7
pytz==2022.7.1
random-password-generator==2.2.0
simple-ddl-parser==0.28.1
six==1.16.0
SQLAlchemy==2.0.4
table-meta==0.1.5
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
